+++ common/src/parsing/timeline.rs
@@ -2,10 +2,12 @@ use crate::parsing::model::Span;
 use crate::parsing::score::{ScalesByName, serialize_scales};
 use crate::pitch::Pitch;
 use num_rational::Ratio;
+use num_traits::ToPrimitive;
 use serde::Serialize;
 use std::borrow::Cow;
 use std::collections::{BTreeMap, BTreeSet};
 use std::sync::Arc;
+use std::{cmp, mem};
 
 #[derive(Serialize)]
 pub struct Timeline<'s> {
@@ -29,7 +31,20 @@ pub struct CsoundGlobalInstrument<'s> {
     pub tail: Ratio<u32>,
 }
 
-#[derive(Serialize, Debug, PartialOrd, PartialEq, Ord, Eq)]
+pub struct TimeBoundaries {
+    pub start_time: Ratio<u32>,
+    pub end_time: Ratio<u32>,
+}
+
+enum TimePosition {
+    Before,
+    After,
+    EndsAtStart,
+    StartsAtEnd,
+    Overlapping,
+}
+
+#[derive(Serialize, Clone, Debug, PartialOrd, PartialEq, Ord, Eq)]
 pub struct TimelineEvent<'s> {
     pub time: Ratio<u32>,
     pub repeat_depth: usize,
@@ -37,6 +52,18 @@ pub struct TimelineEvent<'s> {
     pub data: TimelineData<'s>,
 }
 impl<'s> TimelineEvent<'s> {
+    pub fn end_time(&self) -> Ratio<u32> {
+        let maybe = match &self.data {
+            TimelineData::Note(e) => e.value.pitches.last().map(|x| x.end_time),
+            TimelineData::Tempo(e) => e.end_bpm.as_ref().map(|x| x.time),
+            TimelineData::Dynamic(e) => e.end_level.as_ref().map(|x| x.time),
+            TimelineData::Mark(_) | TimelineData::RepeatStart(_) | TimelineData::RepeatEnd(_) => {
+                None
+            }
+        };
+        maybe.unwrap_or(self.time)
+    }
+
     fn add_or_subtract(v: &mut Ratio<u32>, delta: &Ratio<u32>, subtract: bool) {
         if subtract {
             if delta > v {
@@ -49,8 +76,204 @@ impl<'s> TimelineEvent<'s> {
         }
     }
 
-    pub fn copy_with_time_delta(&self, delta: Ratio<u32>, subtract: bool) -> Self {
+    fn interpolate(
+        value_start: &mut Ratio<u32>,
+        value_end: &mut Ratio<u32>,
+        event_start: &mut Ratio<u32>,
+        event_end: &mut Ratio<u32>,
+        boundaries: &TimeBoundaries,
+    ) -> TimePosition {
+        if *event_start > boundaries.end_time {
+            // This event falls entirely after the end time. We don't care about it.
+            TimePosition::After
+        } else if *event_end == boundaries.start_time {
+            // This event ends exactly at the boundary start time. How it is handled depends on
+            // whether its effect lasts after it's finished, such as tempo.
+            *value_start = *value_end;
+            *event_start = *event_end;
+            TimePosition::EndsAtStart
+        } else if *event_start == boundaries.end_time {
+            // This event starts exactly at the boundary start time. How it is handled depends on
+            // whether we want to know about it, like RepeatEnd or Mark.
+            *value_end = *value_start;
+            *event_end = *event_start;
+            TimePosition::StartsAtEnd
+        } else if *event_end < boundaries.start_time {
+            // This event falls entirely before the start time.
+            TimePosition::Before
+        } else {
+            fn inner(
+                start: Ratio<u32>,
+                end: Ratio<u32>,
+                duration: Ratio<u32>,
+                offset: Ratio<u32>,
+                value: &mut Ratio<u32>,
+            ) {
+                let elapsed_fraction = offset / duration;
+                if start <= end {
+                    let value_delta = end - start;
+                    *value = start + elapsed_fraction * value_delta;
+                } else {
+                    let value_delta = start - end;
+                    *value = start - elapsed_fraction * value_delta;
+                }
+            }
+            let duration = *event_end - *event_start;
+            let mut new_value_start = *value_start;
+            let mut new_value_end = *value_end;
+            let mut new_event_start = *event_start;
+            let mut new_event_end = *event_end;
+            if *event_start < boundaries.start_time {
+                inner(
+                    *value_start,
+                    *value_end,
+                    duration,
+                    boundaries.start_time - *event_start,
+                    &mut new_value_start,
+                );
+                new_event_start = boundaries.start_time;
+            }
+            if *event_end > boundaries.end_time {
+                inner(
+                    *value_start,
+                    *value_end,
+                    duration,
+                    boundaries.end_time - *event_start,
+                    &mut new_value_end,
+                );
+                new_event_end = boundaries.end_time;
+            }
+            *value_start = new_value_start;
+            *value_end = new_value_end;
+            *event_start = new_event_start;
+            *event_end = new_event_end;
+            TimePosition::Overlapping
+        }
+    }
+
+    pub fn copy_with_time_delta(
+        &self,
+        delta: Ratio<u32>,
+        boundaries: Option<&TimeBoundaries>,
+        subtract: bool,
+    ) -> Option<Self> {
         let mut data = self.data.clone();
+        let mut event_start = self.time;
+        if let Some(b) = boundaries {
+            // Filter out or adjust an event based on where it falls within the region of time we
+            // are interested in. At the end, if the event survives, the event start time is moved
+            // forward to the time boundary.
+            match &mut data {
+                TimelineData::Tempo(e) => {
+                    match &mut e.end_bpm {
+                        // This is a gradual tempo change event. Interpolate to set the range
+                        // based on where we are in the tempo change.
+                        Some(end_bpm) => {
+                            let time_pos = Self::interpolate(
+                                &mut e.bpm,
+                                &mut end_bpm.item,
+                                &mut event_start,
+                                &mut end_bpm.time,
+                                b,
+                            );
+                            match time_pos {
+                                TimePosition::Before | TimePosition::EndsAtStart => {
+                                    // If the tempo change finished at or before the start time,
+                                    // treat it as an instantaneous tempo event for the end BPM.
+                                    e.bpm = end_bpm.item;
+                                    e.end_bpm = None;
+                                }
+                                TimePosition::After | TimePosition::StartsAtEnd => return None,
+                                TimePosition::Overlapping => {}
+                            }
+                        }
+                        None => {
+                            // Keep any instantaneous tempo event that happens any time before the
+                            // end time.
+                            if event_start >= b.end_time {
+                                return None;
+                            }
+                        }
+                    }
+                }
+                TimelineData::Dynamic(e) => {
+                    if let Some(end_level) = &mut e.end_level {
+                        // We are part way through a dynamic change. Interpolate and then
+                        // force back to u8.
+                        let mut start_value: Ratio<u32> = Ratio::from_integer(e.start_level.into());
+                        let mut end_value: Ratio<u32> = Ratio::from_integer(end_level.item.into());
+                        if !matches!(
+                            Self::interpolate(
+                                &mut start_value,
+                                &mut end_value,
+                                &mut event_start,
+                                &mut end_level.time,
+                                b,
+                            ),
+                            TimePosition::Overlapping
+                        ) {
+                            return None;
+                        }
+                        // Interpolated values can only ever move toward their midpoint, so
+                        // rounding the resulted interpolated values will always result in
+                        // values that are within the valid u8 range.
+                        e.start_level = start_value.round().to_u8().unwrap();
+                        end_level.item = end_value.round().to_u8().unwrap();
+                    }
+                }
+                TimelineData::Note(e) => {
+                    let pitches: Vec<PitchChange> = mem::take(&mut e.value.pitches);
+                    for mut p in pitches {
+                        // Interpolating pitch values is tricky, so use proxies to compute the
+                        // argument to Pitch::interpolate.
+                        let mut value_start = Ratio::from_integer(0);
+                        let mut value_end = Ratio::from_integer(1);
+                        if !matches!(
+                            Self::interpolate(
+                                &mut value_start,
+                                &mut value_end,
+                                &mut p.start_time,
+                                &mut p.end_time,
+                                b,
+                            ),
+                            TimePosition::Overlapping
+                        ) {
+                            continue;
+                        }
+                        if let Some(end_pitch) = &mut p.end_pitch {
+                            let new_start_pitch =
+                                Pitch::interpolate(&p.start_pitch, end_pitch, value_start);
+                            let new_end_pitch =
+                                Pitch::interpolate(&p.start_pitch, end_pitch, value_end);
+                            p.start_pitch = new_start_pitch;
+                            *end_pitch = new_end_pitch;
+                        }
+                        e.value.pitches.push(p);
+                    }
+                    if e.value.pitches.is_empty() {
+                        return None;
+                    }
+                }
+                TimelineData::Mark(_) => {
+                    // Mark events are instantaneous, and we only care about them if they are in
+                    // the time range, not at the boundaries.
+                    if event_start <= b.start_time || event_start >= b.end_time {
+                        return None;
+                    }
+                }
+                TimelineData::RepeatStart(_) | TimelineData::RepeatEnd(_) => {
+                    // These events are okay to keep when they appear at the boundaries.
+                    if event_start < b.start_time || event_start > b.end_time {
+                        return None;
+                    }
+                }
+            }
+            // Static or single-point-of-time events should be adjusted to be within range.
+            // This will be harmlessly redundant for events that were interpolated if the
+            // actual event start time was passed to interpolate, but that is not always the
+            // case.
+            event_start = cmp::max(event_start, b.start_time);
+        }
         match &mut data {
             TimelineData::Tempo(e) => {
                 if let Some(x) = e.end_bpm.as_mut() {
@@ -70,18 +293,20 @@ impl<'s> TimelineEvent<'s> {
             }
             TimelineData::Mark(_) | TimelineData::RepeatStart(_) | TimelineData::RepeatEnd(_) => {}
         };
-        let mut new_time = self.time;
+
+        let mut new_time = event_start;
         Self::add_or_subtract(&mut new_time, &delta, subtract);
-        Self {
+        Some(Self {
             time: new_time,
             repeat_depth: self.repeat_depth,
             span: self.span,
             data,
-        }
+        })
     }
 
     pub fn copy_for_repeat(&self, delta: Ratio<u32>) -> Self {
-        let mut event = self.copy_with_time_delta(delta, false);
+        // copy_with_time_delta always returns Some when time_boundaries is None.
+        let mut event = self.copy_with_time_delta(delta, None, false).unwrap();
         event.repeat_depth += 1;
         event
     }
@@ -213,5 +438,102 @@ mod tests {
             iter.cloned().collect::<Vec<_>>(),
             [(3, 1), (4, 1), (4, 2), (5, 1)]
         );
+    }
+
+    #[test]
+    fn test_add_or_subtract() {
+        // add_or_subtract is thoroughly tested through other means, but the way event filtering
+        // is implemented prevents the boundary condition of subtracting too much from ever
+        // happening organically.
+        let mut v = Ratio::from_integer(3);
+        TimelineEvent::add_or_subtract(&mut v, &Ratio::from_integer(1), true);
+        assert_eq!(v, Ratio::from_integer(2));
+        TimelineEvent::add_or_subtract(&mut v, &Ratio::from_integer(5), true);
+        assert_eq!(v, Ratio::from_integer(0));
+    }
+
+    #[test]
+    fn test_interpolate() {
+        fn r(n: u32) -> Ratio<u32> {
+            Ratio::from_integer(n)
+        }
+        fn tb(start_time: u32, end_time: u32) -> TimeBoundaries {
+            TimeBoundaries {
+                start_time: r(start_time),
+                end_time: r(end_time),
+            }
+        }
+
+        // These need to be mut. The allow statements are working around a RustRover false positive.
+        // as of RustRover 2026.1.
+        // https://youtrack.jetbrains.com/issue/RUST-20121/false-positive-unused-mut-with-macrorules
+        #[allow(unused_mut)]
+        let mut value_start;
+        #[allow(unused_mut)]
+        let mut value_end;
+        #[allow(unused_mut)]
+        let mut event_start;
+        #[allow(unused_mut)]
+        let mut event_end;
+        macro_rules! interpolate {
+            ($v1:expr, $v2:expr, $e1:expr, $e2:expr, $b1:expr, $b2:expr) => {{
+                value_start = r($v1);
+                value_end = r($v2);
+                event_start = r($e1);
+                event_end = r($e2);
+                TimelineEvent::interpolate(
+                    &mut value_start,
+                    &mut value_end,
+                    &mut event_start,
+                    &mut event_end,
+                    &tb($b1, $b2),
+                )
+            }};
+        }
+        assert!(matches!(
+            interpolate!(60, 120, 0, 12, 3, 15),
+            TimePosition::Overlapping
+        ));
+        assert_eq!(event_start, r(3));
+        assert_eq!(event_end, r(12));
+        assert_eq!(value_start, r(75));
+        assert_eq!(value_end, r(120));
+
+        assert!(matches!(
+            interpolate!(60, 120, 0, 12, 9, 15),
+            TimePosition::Overlapping
+        ));
+        assert_eq!(event_start, r(9));
+        assert_eq!(event_end, r(12));
+        assert_eq!(value_start, r(105));
+        assert_eq!(value_end, r(120));
+
+        assert!(matches!(
+            interpolate!(60, 120, 0, 12, 12, 15),
+            TimePosition::EndsAtStart
+        ));
+        assert_eq!(event_start, r(12));
+        assert_eq!(event_end, r(12));
+        assert_eq!(value_start, r(120));
+        assert_eq!(value_end, r(120));
+
+        assert!(matches!(
+            interpolate!(60, 120, 3, 12, 0, 3),
+            TimePosition::StartsAtEnd
+        ));
+        assert_eq!(event_start, r(3));
+        assert_eq!(event_end, r(3));
+        assert_eq!(value_start, r(60));
+        assert_eq!(value_end, r(60));
+
+        assert!(matches!(
+            interpolate!(60, 120, 0, 12, 14, 15),
+            TimePosition::Before
+        ));
+
+        assert!(matches!(
+            interpolate!(60, 120, 3, 12, 0, 2),
+            TimePosition::After
+        ));
     }
 }